requirements.txt
flask
pandas
numpy
requests
scikit-learn
sentence-transformers
faiss-cpu
python-dotenv
google-generativeai
torch